requirements.txt
fastapi
uvicorn
jinja2
python-multipart
openai
python-dotenv
google-generativeai